requirements.txt
torch~=2.0.0
numpy~=1.23.5
pillow~=9.4.0
torchvision~=0.15.1
scipy~=1.10.0
imgaug~=0.4.0
matplotlib~=3.7.1
tqdm~=4.65.0
opencv-python~=4.8.0.74